+++ module/ldbc-connector/shared/src/main/scala/ldbc/connector/net/protocol/ClientPreparedStatement.scala
@@ -56,7 +56,7 @@ case class ClientPreparedStatement[F[_]: Temporal: Exchange: Tracer](
   resultSetType:        Int = ResultSet.TYPE_FORWARD_ONLY,
   resultSetConcurrency: Int = ResultSet.CONCUR_READ_ONLY
 )(using ev: MonadError[F, Throwable])
-  extends PreparedStatementImpl.Share[F]:
+  extends SharedPreparedStatement[F]:
 
   private val attributes = protocol.initialPacket.attributes ++ List(
     Attribute("type", "Client PreparedStatement"),
